All fixed after abstract PaveralItem changes
All tested, added Todos for newly discovered unrelated bugs. Crafting is ready!
