+++ src/main/java/me/powerspieler/paveral/Paveral.java
@@ -13,8 +13,8 @@ import me.powerspieler.paveral.forge.ForgeListener;
 import me.powerspieler.paveral.forming_altar.AwakeAltar;
 import me.powerspieler.paveral.forming_altar.FormingListeners;
 import me.powerspieler.paveral.items.*;
-import me.powerspieler.paveral.items.enchanced.Channeling;
-import me.powerspieler.paveral.items.enchanced.Knockback;
+import me.powerspieler.paveral.items.enhanced.Channeling;
+import me.powerspieler.paveral.items.enhanced.Knockback;
 import me.powerspieler.paveral.items.musicpack.LumberjacksBass;
 import me.powerspieler.paveral.items.musicpack.PianoSword;
 import me.powerspieler.paveral.items.musicpack.ScytheOfHarmony;
