+++ src/main/java/me/powerspieler/paveral/discovery/tutorial/TechBook.java
@@ -21,7 +21,7 @@ public class TechBook extends DiscoveryBook implements Formable {
         pages.add(GsonComponentSerializer.gson().deserialize("[{\"text\":\"I am glad you made it, once again!\nI hope you appreciate the effort I put into the advancements.\n\nAnyway, you may have noticed that all craftable items are based on the same theme so far:\n\n\",\"color\":\"dark_purple\"},{\"text\":\" ~< Magic >~\",\"color\":\"light_purple\"}]"));
         pages.add(GsonComponentSerializer.gson().deserialize("[{\"text\":\"This is the moment of introducing another:\n\n\",\"color\":\"dark_purple\"},{\"text\":\"~<Technics>~\n\n\",\"color\":\"dark_blue\"},{\"text\":\"\\\"Forging\\\" technical items also requires a multiblock structure like the process of forming on the altar.\n\nAs always the materials are listed on the next page.\",\"color\":\"dark_purple\"}]"));
         pages.add(GsonComponentSerializer.gson().deserialize("[{\"text\":\"Materials:\n\",\"color\":\"light_purple\"},{\"text\":\"8x Polished Basalt\n8x Cobbled Deepslate\n    Slab\n4x Cauldron\n4x Hopper\n4x Polished\n    Blackstone Brick\n    Wall\n4x Cobbled Deepslate\n4x Polished\n    Blackstone Bricks\n1x Anvil\n1x Lodestone\",\"color\":\"green\"}]"));
-        pages.add(GsonComponentSerializer.gson().deserialize("[{\"text\":\"Layer 1:\n\",\"color\":\"dark_purple\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"black\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Blackstone Bricks\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate Slab\"}]}},{\"text\":\"⬛\",\"color\":\"dark_gray\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Basalt (Facing towards mid)\"}]}},{\"text\":\"⬛\",\"color\":\"blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate Slab\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate Slab\"}]}},{\"text\":\"⬛\",\"color\":\"dark_blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate\"}]}},{\"text\":\"⬛\",\"color\":\"dark_gray\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Basalt (Facing towards mid)\"}]}},{\"text\":\"⬛\",\"color\":\"dark_blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate\"}]}},{\"text\":\"⬛\",\"color\":\"blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate Slab\"}]}},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"black\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Blackstone Blocks\"}]}},{\"text\":\"⬛\",\"color\":\"dark_gray\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Basalt (Facing towards mid)\"}]}},{\"text\":\"⬛\",\"color\":\"dark_gray\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Basalt (Facing towards mid)\"}]}},{\"text\":\"⬛\",\"color\":\"gray\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Lodestone\"}]}},{\"text\":\"⬛\",\"color\":\"dark_gray\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Basalt (Facing towards mid)\"}]}},{\"text\":\"⬛\",\"color\":\"dark_gray\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Basalt (Facing towards mid)\"}]}},{\"text\":\"⬛\n\",\"color\":\"black\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Blackstone Blocks\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate Slab\"}]}},{\"text\":\"⬛\",\"color\":\"dark_blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate\"}]}},{\"text\":\"⬛\",\"color\":\"dark_gray\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Basalt (Facing towards mid)\"}]}},{\"text\":\"⬛\",\"color\":\"dark_blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate\"}]}},{\"text\":\"⬛\",\"color\":\"blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate Slab\"}]}},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate Slab\"}]}},{\"text\":\"⬛\",\"color\":\"dark_gray\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Basalt (Facing towards mid)\"}]}},{\"text\":\"⬛\",\"color\":\"blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate Slab\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"black\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Blackstone Blocks\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"\\nHover for Details\",\"color\":\"dark_purple\"}]"));
+        pages.add(GsonComponentSerializer.gson().deserialize("[{\"text\":\"Layer 1:\n\",\"color\":\"dark_purple\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"black\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Blackstone Bricks\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate Slab\"}]}},{\"text\":\"⬛\",\"color\":\"dark_gray\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Basalt (Facing towards mid)\"}]}},{\"text\":\"⬛\",\"color\":\"blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate Slab\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate Slab\"}]}},{\"text\":\"⬛\",\"color\":\"dark_blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate\"}]}},{\"text\":\"⬛\",\"color\":\"dark_gray\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Basalt (Facing towards mid)\"}]}},{\"text\":\"⬛\",\"color\":\"dark_blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate\"}]}},{\"text\":\"⬛\",\"color\":\"blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate Slab\"}]}},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"black\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Blackstone Bricks\"}]}},{\"text\":\"⬛\",\"color\":\"dark_gray\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Basalt (Facing towards mid)\"}]}},{\"text\":\"⬛\",\"color\":\"dark_gray\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Basalt (Facing towards mid)\"}]}},{\"text\":\"⬛\",\"color\":\"gray\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Lodestone\"}]}},{\"text\":\"⬛\",\"color\":\"dark_gray\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Basalt (Facing towards mid)\"}]}},{\"text\":\"⬛\",\"color\":\"dark_gray\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Basalt (Facing towards mid)\"}]}},{\"text\":\"⬛\n\",\"color\":\"black\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Blackstone Bricks\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate Slab\"}]}},{\"text\":\"⬛\",\"color\":\"dark_blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate\"}]}},{\"text\":\"⬛\",\"color\":\"dark_gray\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Basalt (Facing towards mid)\"}]}},{\"text\":\"⬛\",\"color\":\"dark_blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate\"}]}},{\"text\":\"⬛\",\"color\":\"blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate Slab\"}]}},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate Slab\"}]}},{\"text\":\"⬛\",\"color\":\"dark_gray\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Basalt (Facing towards mid)\"}]}},{\"text\":\"⬛\",\"color\":\"blue\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cobbled Deepslate Slab\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"black\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Blackstone Bricks\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"\\nHover for Details\",\"color\":\"dark_purple\"}]"));
         pages.add(GsonComponentSerializer.gson().deserialize("[{\"text\":\"Layer 2:\n\",\"color\":\"dark_purple\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"dark_purple\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Blackstone Brick Wall\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"dark_purple\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Blackstone Brick Wall\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"dark_gray\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Anvil\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"dark_purple\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Blackstone Brick Wall\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"dark_purple\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Polished Blackstone Brick Wall\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"\\nHover for Details\",\"color\":\"dark_purple\"}]"));
         pages.add(GsonComponentSerializer.gson().deserialize("[{\"text\":\"Layer 3:\n\",\"color\":\"dark_purple\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"light_purple\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Hopper\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"light_purple\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Hopper\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"light_purple\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Hopper\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"light_purple\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Hopper\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"\\nHover for Details\",\"color\":\"dark_purple\"}]"));
         pages.add(GsonComponentSerializer.gson().deserialize("[{\"text\":\"Layer 4:\n\",\"color\":\"dark_purple\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"#FF6A00\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cauldron\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"#FF6A00\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cauldron\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"#FF6A00\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cauldron\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"#FF6A00\",\"hoverEvent\":{\"action\":\"show_text\",\"contents\":[{\"text\":\"Cauldron\"}]}},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\",\"color\":\"white\"},{\"text\":\"⬛\n\",\"color\":\"white\"},{\"text\":\"\\nHover for Details\",\"color\":\"dark_purple\"}]"));
