+++ src/main/java/me/powerspieler/paveral/items/ItemHoldingController.java
@@ -9,6 +9,9 @@ import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.entity.EntityPickupItemEvent;
 import org.bukkit.event.inventory.InventoryClickEvent;
+import org.bukkit.event.inventory.InventoryDragEvent;
+import org.bukkit.event.inventory.InventoryEvent;
+import org.bukkit.event.inventory.InventoryMoveItemEvent;
 import org.bukkit.event.player.PlayerDropItemEvent;
 import org.bukkit.event.player.PlayerInteractAtEntityEvent;
 import org.bukkit.event.player.PlayerItemHeldEvent;
@@ -17,7 +20,7 @@ import org.bukkit.inventory.ItemStack;
 import org.bukkit.persistence.PersistentDataType;
 import org.bukkit.scheduler.BukkitRunnable;
 
-public class ItemHoldingController implements Listener {
+public class ItemHoldingController implements Listener { //TODO InventoryDragEvent , Chest Item Transfer (?), etc.. considering on every isHoldingPaveralItem call recheck
     @EventHandler
     public void onItemSwitch(PlayerItemHeldEvent event){
         Player player = event.getPlayer();
