+++ src/main/java/me/powerspieler/paveral/items/enhanced/Channeling.java
@@ -1,9 +1,10 @@
-package me.powerspieler.paveral.items.enchanced;
+package me.powerspieler.paveral.items.enhanced;
 
 import me.powerspieler.paveral.crafting.EnchantmentIngredient;
 import me.powerspieler.paveral.crafting.PaveralRecipe;
 import me.powerspieler.paveral.crafting.StandardIngredient;
 import me.powerspieler.paveral.items.Dismantable;
+import me.powerspieler.paveral.items.LightningRod;
 import me.powerspieler.paveral.items.PaveralItem;
 import me.powerspieler.paveral.util.Constant;
 import net.kyori.adventure.text.Component;
@@ -16,12 +17,11 @@ import org.bukkit.event.Listener;
 import org.bukkit.event.inventory.PrepareAnvilEvent;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.EnchantmentStorageMeta;
-import org.bukkit.inventory.meta.ItemMeta;
 import org.bukkit.persistence.PersistentDataType;
 
 import java.util.*;
 
-public class Knockback extends PaveralItem implements Listener, Dismantable {
+public class Channeling extends PaveralItem implements Listener, Dismantable {
     private static Component itemName(){
         return Component.text("Enhanced Book", NamedTextColor.LIGHT_PURPLE)
                 .decoration(TextDecoration.ITALIC, false);
@@ -31,20 +31,20 @@ public class Knockback extends PaveralItem implements Listener, Dismantable {
         List<Component> lore = new ArrayList<>();
         lore.add(Component.text("Compatible with: ")
                 .decoration(TextDecoration.ITALIC, false)
-                .append(Component.text("Stick", NamedTextColor.YELLOW)
+                .append(Component.text("Trident", NamedTextColor.YELLOW)
                         .decoration(TextDecoration.ITALIC, false)));
         return lore;
     }
 
-    public Knockback() {
-        super(Material.ENCHANTED_BOOK, 0, Constant.ITEMTYPE, "enhanced_knockback", itemName(), lore());
+    public Channeling() {
+        super(Material.ENCHANTED_BOOK, 0, Constant.ITEMTYPE, "enhanced_channeling", itemName(), lore());
     }
 
     @Override
     protected ItemStack build() {
         ItemStack item = super.build();
         EnchantmentStorageMeta itemMeta = (EnchantmentStorageMeta) item.getItemMeta();
-        itemMeta.addStoredEnchant(Enchantment.KNOCKBACK, 5, true);
+        itemMeta.addStoredEnchant(Enchantment.CHANNELING, 10, true);
         item.setItemMeta(itemMeta);
         return item;
     }
@@ -52,7 +52,7 @@ public class Knockback extends PaveralItem implements Listener, Dismantable {
     @Override
     public PaveralRecipe recipe() {
         Map<Enchantment, Integer> enchantments = new HashMap<>();
-        enchantments.put(Enchantment.KNOCKBACK, 2);
+        enchantments.put(Enchantment.CHANNELING, 1);
 
         Set<StandardIngredient> ingredients = new HashSet<>();
         ingredients.add(new EnchantmentIngredient(enchantments));
@@ -71,21 +71,8 @@ public class Knockback extends PaveralItem implements Listener, Dismantable {
 
     @EventHandler
     private void onAnvilUse(PrepareAnvilEvent event){
-        if(event.getInventory().getFirstItem() != null && event.getInventory().getFirstItem().getType() == Material.STICK && event.getInventory().getSecondItem() != null && Objects.equals(event.getInventory().getSecondItem().getItemMeta().getPersistentDataContainer().get(Constant.ITEMTYPE, PersistentDataType.STRING), keyString)){
-            ItemStack item = new ItemStack(Material.STICK);
-            ItemMeta itemmeta = item.getItemMeta();
-            itemmeta.getPersistentDataContainer().set(Constant.ITEMTYPE, PersistentDataType.STRING, "bonk");
-            itemmeta.addEnchant(Enchantment.KNOCKBACK, 5, true);
-            itemmeta.itemName(Component.text("Bonk", NamedTextColor.RED)
-                    .decoration(TextDecoration.ITALIC, false));
-            List<Component> lore = new ArrayList<>();
-            lore.add(Component.text("Hornyjail certified", NamedTextColor.GOLD)
-                    .decoration(TextDecoration.ITALIC, false));
-            itemmeta.lore(lore);
-            item.setItemMeta(itemmeta);
-            event.getView().setRepairCost(0);
-            event.setResult(item);
-
+        if(event.getInventory().getFirstItem() != null && event.getInventory().getFirstItem().getType() == Material.TRIDENT && event.getInventory().getSecondItem() != null && Objects.equals(event.getInventory().getSecondItem().getItemMeta().getPersistentDataContainer().get(Constant.ITEMTYPE, PersistentDataType.STRING), keyString)){
+            event.setResult(new LightningRod().build());
         }
     }
 }
